+++ src/main/java/net/geforcemods/securitycraft/blockentities/ScannerDoorBlockEntity.java
@@ -4,6 +4,7 @@ import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.IViewActivated;
 import net.geforcemods.securitycraft.api.Option;
+import net.geforcemods.securitycraft.api.Option.BooleanOption;
 import net.geforcemods.securitycraft.api.Option.DoubleOption;
 import net.geforcemods.securitycraft.api.Owner;
 import net.geforcemods.securitycraft.blocks.ScannerDoorBlock;
@@ -24,6 +25,7 @@ import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.common.util.Constants;
 
 public class ScannerDoorBlockEntity extends SpecialDoorBlockEntity implements IViewActivated {
+	private BooleanOption sendMessage = new BooleanOption("sendMessage", true);
 	private DoubleOption maximumDistance = new DoubleOption(this::getPos, "maximumDistance", 5.0D, 0.1D, 25.0D, 0.1D, true) {
 		@Override
 		public String getKey(Block block) {
@@ -111,6 +113,10 @@ public class ScannerDoorBlockEntity extends SpecialDoorBlockEntity implements IV
 		return new Option[] {
 				sendMessage, signalLength, disabled, maximumDistance
 		};
+	}
+
+	public boolean sendsMessages() {
+		return sendMessage.get();
 	}
 
 	@Override