+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadFurnaceBlock.java
@@ -108,25 +108,23 @@ public class KeypadFurnaceBlock extends DisguisableBlock {
 	@Override
 	public boolean onBlockActivated(World world, BlockPos pos, IBlockState state, EntityPlayer player, EnumHand hand, EnumFacing side, float hitX, float hitY, float hitZ) {
 		if (!world.isRemote) {
-			KeypadFurnaceBlockEntity te = (KeypadFurnaceBlockEntity) world.getTileEntity(pos);
+			KeypadFurnaceBlockEntity be = (KeypadFurnaceBlockEntity) world.getTileEntity(pos);
 
-			if (te.isDisabled())
+			if (be.isDisabled())
 				player.sendStatusMessage(Utils.localize("gui.securitycraft:scManual.disabled"), true);
-			else if (te.verifyPasscodeSet(world, pos, te, player)) {
-				if (te.isDenied(player)) {
-					if (te.sendsMessages())
+			else if (be.verifyPasscodeSet(world, pos, be, player)) {
+				if (be.isDenied(player)) {
+					if (be.sendsDenylistMessage())
 						PlayerUtils.sendMessageToPlayer(player, Utils.localize(getTranslationKey() + ".name"), Utils.localize("messages.securitycraft:module.onDenylist"), TextFormatting.RED);
-
-					return true;
 				}
-				else if (te.isAllowed(player)) {
-					if (te.sendsMessages())
+				else if (be.isAllowed(player)) {
+					if (be.sendsAllowlistMessage())
 						PlayerUtils.sendMessageToPlayer(player, Utils.localize(getTranslationKey() + ".name"), Utils.localize("messages.securitycraft:module.onAllowlist"), TextFormatting.GREEN);
 
 					activate(state, world, pos, player);
 				}
 				else if (player.getHeldItem(hand).getItem() != SCContent.codebreaker)
-					te.openPasscodeGUI(world, pos, player);
+					be.openPasscodeGUI(world, pos, player);
 			}
 		}